test: move register validation tests to dedicated suite
